use pandas; csv output of user attr
